+++ src/editor/core/plugins/link/schema.js
@@ -6,6 +6,7 @@
  */
 
 import {validateHref, DEFAULT_LINK_REL} from "../../util/linkUtil";
+import {filterFileUrl} from "../../humhub-bridge";
 
 const schema = {
     marks: {
@@ -37,17 +38,16 @@ const schema = {
                 }],
             parseMarkdown: {
                 mark: "link", getAttrs: function (tok) {
-                    let href = (window.humhub) ? humhub.modules.file.filterFileUrl(tok.attrGet("href")).url : tok.attrGet("href");
-                    let fileGuid = (window.humhub) ? humhub.modules.file.filterFileUrl(tok.attrGet("href")).guid : null;
+                    let {url, guid} = filterFileUrl(tok.attrGet("href"));
 
-                    if (!validateHref(href))  {
-                        href = '#';
+                    if (!validateHref(url))  {
+                        url = '#';
                     }
 
                     return ({
-                        href: href,
+                        href: url,
                         title: tok.attrGet("title") || null,
-                        fileGuid: fileGuid
+                        fileGuid: guid
                     });
                 }
             },